test: add test data, rename test files for jest
