+++ test/fixtures/connection.fixture.ts
@@ -1,10 +1,12 @@
 import Knex from 'knex';
-import { Tables } from 'src/database/tables.constants';
+import { Tables } from '../../src/database/tables.constants';
 import * as knexConfigs from '../../src/database/knexfile';
+import { seed as usersSeed } from '../../src/database/seeds/test_data/user.seed';
 
 /** Create a new knex instance for testing */
 export const getTestConnection = async () => {
   const kn = knexFatory();
+  return await seedDb(await setupSchema(kn));
 };
 
 /** Factory for creating knex connection instance */
@@ -38,7 +40,29 @@ export const truncateDb = async (kn: Knex): Promise<Knex> => {
 };
 
 /** Executes the given seed functions aginast the given knex connection. */
-// export const seedDb = async(
-//     kn: Knex,
-//     seed: ((kn: Knex) => Promise<void>)[] = [user]
-// )
+export const seedDb = async (
+  kn: Knex,
+  seeds: ((kn: Knex) => Promise<void>)[] = [usersSeed],
+): Promise<Knex> => {
+  // Ensure usersSeed isn't added twice
+  const index = seeds.indexOf(usersSeed);
+  if (index < 0) {
+    seeds = [...seeds, usersSeed];
+  }
+  await Promise.all(
+    seeds.map(seed => {
+      return seed(kn);
+    }),
+  );
+  return kn;
+};
+
+/** Truncates and reseeds the db for the given connection */
+export const refreshDb = async (
+  kn: Knex,
+  seeds: ((kn: Knex) => Promise<void>)[] = [],
+) => {
+  await truncateDb(kn);
+  await seedDb(kn, seeds);
+  return kn;
+};
